Move Derivative struct into its own module
